requirements.txt
# requirements.txt
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
psutil>=5.9.0
icrawler>=0.6.0
gtts>=2.3.0
googletrans==4.0.0-rc1
requests>=2.28.0